update location of extension
