+++ internal/util/extensions_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v20260728
+package util
 
 import (
 	"context"
